+++ primitives/parent.hh
@@ -14,7 +14,7 @@ class dcParent : public dcObject
         void draw(void);
         void handleKeyPress(char);
         void handleKeyRelease(char);
-        void handleMousePress(float, float);
+        void handleMousePress(double, double);
         void handleMouseRelease(void);
         void handleBezelPress(int);
         void handleBezelRelease(int);
